Refactor core to return lazy seqs which block next
Recording endpoint now returns a lazy sequence of requests. As one
iterates through the lazy seq, an infinite seq of request futures is
also iterated.  When the next element of the lazy seq is requested, the
next future is dereferenced with a timeout. If the next future cannot
be dereferenced within the timeout span the lazy-seq is terminated.

This allows one to use all of the standard sequence functions.