Wait for slow handlers before reporting request
This helps, but does not remove, a cause of flakiness in client code.
Imagine a test running in a foreground thread. Suppose it starts a
system which runs a server A in the background. Then the test triggers
another background thread B which makes a request to the
standalone-server C. Also suppose the test correctly waits for the
standalone-server C to finish responding to the request. Then requests
to A can still be flaky: if the background thread B is slow it might not
finish what it's doing before the test calls the server A to see what B
did.

This patch fixes flakiness where C is slow to respond, but does not help
when B is slow to process the response.