Split recording of request and response traffic
The sentiment in #0a2be40 was correct, some tests want to wait for slow
handlers to finish before proceeding.

However, other tests want to know only that a request has been made to
a slow handler. They do not care whether the handler ever responds, and
may even expect the client of the handler to timeout.

This patch lets tests record either when requests start or when
responses finish.